lavfi/scale_qsv: re-use VPPContext for scale_qsv filter
All features are implemented in vpp_qsv filter, scale_qsv can be taken
as a special case of vpp_qsv filter now, we re-use VPPContext with a
different option arrary and pixel formats
